+++ plotnine/coords/coord_cartesian.py
@@ -1,5 +1,6 @@
-import types
-from mizani.bounds import expand_range_distinct, squish_infinite
+from types import SimpleNamespace as NS
+from mizani.bounds import squish_infinite
+from mizani.transforms import identity_trans
 
 from .coord import coord, dist_euclidean
 from ..positions.position import transform_position
@@ -26,15 +27,15 @@ class coord_cartesian(coord):
     is_linear = True
 
     def __init__(self, xlim=None, ylim=None, expand=True):
-        self.limits = types.SimpleNamespace(xlim=xlim, ylim=ylim)
+        self.limits = NS(x=xlim, y=ylim)
         self.expand = expand
 
     def transform(self, data, panel_params, munch=False):
         def squish_infinite_x(data):
-            return squish_infinite(data, range=panel_params['x_range'])
+            return squish_infinite(data, range=panel_params.x.range)
 
         def squish_infinite_y(data):
-            return squish_infinite(data, range=panel_params['y_range'])
+            return squish_infinite(data, range=panel_params.y.range)
 
         return transform_position(data, squish_infinite_x, squish_infinite_y)
 
@@ -42,43 +43,19 @@ class coord_cartesian(coord):
         """
         Compute the range and break information for the panel
         """
+        def get_view_limits(scale, coord_limits):
+            expansion = scale.default_expansion(expand=self.expand)
+            ranges = scale.expand_limits(
+                scale.limits, expansion, coord_limits, identity_trans)
+            vs = scale.view(limits=coord_limits, range=ranges.range)
+            return vs
 
-        def train(scale, limits, name):
-            """
-            Train a single coordinate axis
-            """
-            # Which axis are we dealing with
-            name = scale.aesthetics[0]
-
-            if self.expand:
-                expand = self.expand_default(scale)
-            else:
-                expand = (0, 0, 0, 0)
-
-            if limits is None:
-                rangee = scale.dimension(expand)
-            else:
-                rangee = scale.transform(limits)
-                rangee = expand_range_distinct(rangee, expand)
-
-            out = scale.break_info(rangee)
-            # This is where
-            # x_major, x_labels, x_minor, ...
-            # range keys are created
-            for key in list(out.keys()):
-                new_key = '{}_{}'.format(name, key)
-                out[new_key] = out.pop(key)
-            return out
-
-        out = dict(
-            scales=types.SimpleNamespace(x=scale_x, y=scale_y),
-            **train(scale_x, self.limits.xlim, 'x'),
-            **train(scale_y, self.limits.ylim, 'y')
-        )
+        out = NS(x=get_view_limits(scale_x, self.limits.x),
+                 y=get_view_limits(scale_y, self.limits.y))
         return out
 
     @staticmethod
     def distance(x, y, panel_params):
-        max_dist = dist_euclidean(panel_params['x_range'],
-                                  panel_params['y_range'])[0]
+        max_dist = dist_euclidean(panel_params.x.range,
+                                  panel_params.y.range)[0]
         return dist_euclidean(x, y) / max_dist